+++ Tests/Functional/ViewHelpers/AdminPanel/PropertyValueViewHelperTest.php
@@ -18,6 +18,7 @@ use Brotkrueml\Schema\ViewHelpers\AdminPanel\PropertyValueViewHelper;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
+use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Fluid\Core\Rendering\RenderingContextFactory;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 use TYPO3Fluid\Fluid\View\TemplateView;
@@ -45,6 +46,12 @@ final class PropertyValueViewHelperTest extends FunctionalTestCase
         string $template,
         string $expected,
     ): void {
+        $typo3Version = (new Typo3Version())->getMajorVersion();
+        if ($typo3Version >= 14) {
+            // @todo Find solution to test it
+            self::markTestSkipped('Dynamic output with timestamp on assets');
+        }
+
         $context = $this->get(RenderingContextFactory::class)->create();
         $context->getTemplatePaths()->setTemplateSource($template);
 
@@ -284,6 +291,12 @@ EXPECTED,
     #[Test]
     public function additionalManualsAreRenderedCorrectly(): void
     {
+        $typo3Version = (new Typo3Version())->getMajorVersion();
+        if ($typo3Version >= 14) {
+            // @todo Find solution to test it
+            self::markTestSkipped('Dynamic output with timestamp on assets');
+        }
+
         $typeRegistry = $this->get(TypeRegistry::class);
         $typeRegistry->addType('Thing', Thing::class);
         $typeRegistry->addManualForType('Thing', [Publisher::Google, 'Some link', 'https://example.org/Thing']);